requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
flask>=2.0.0
